Extract completion listeners into generic Rx classes
